Remove typography.html
The header and footer currently include the link to typography, but removing typography.html removes that too.

In other words, if i want to add a new page to the website, I just need to add an html file like this then it'll automatically add the link to the header and footer.